update for latest antlr

+++ runtime/Rust/src/atn.rs
@@ -14,7 +14,7 @@ use crate::token_factory::CommonTokenFactory;
 use crate::transition::RuleTransition;
 use std::fmt::{Debug, Formatter};
 
-pub const INVALID_ALT: isize = 0;
+pub const INVALID_ALT: i32 = 0;
 
 /// Augmented Transition Network
 ///
@@ -28,7 +28,7 @@ pub struct ATN {
 
     pub(crate) lexer_actions: Vec<LexerAction>,
 
-    pub max_token_type: isize,
+    pub max_token_type: i32,
 
     pub mode_name_to_start_state: HashMap<String, ATNStateRef>,
 
@@ -38,7 +38,7 @@ pub struct ATN {
 
     pub rule_to_stop_state: Vec<ATNStateRef>,
 
-    pub rule_to_token_type: Vec<isize>,
+    pub rule_to_token_type: Vec<i32>,
 
     pub states: Vec<Box<dyn ATNState>>,
 }
@@ -55,7 +55,7 @@ impl Debug for ATN {
 }
 
 impl ATN {
-    pub(crate) fn new_atn(grammar_type: ATNType, max_token_type: isize) -> ATN {
+    pub(crate) fn new_atn(grammar_type: ATNType, max_token_type: i32) -> ATN {
         ATN {
             decision_to_state: Vec::new(),
             grammar_type,
@@ -94,16 +94,16 @@ impl ATN {
     }
 
     pub(crate) fn add_state(&mut self, state: Box<dyn ATNState>) {
-        debug_assert_eq!(state.get_state_number(), self.states.len());
+        debug_assert_eq!(state.get_state_number() as usize, self.states.len());
         self.states.push(state)
     }
 
     // fn remove_state(&self, _state: ATNStateRef) { unimplemented!() }
 
-    // fn define_decision_state(&self, _s: ATNStateRef) -> isize { unimplemented!() }
+    // fn define_decision_state(&self, _s: ATNStateRef) -> i32 { unimplemented!() }
 
-    pub fn get_decision_state(&self, decision: usize) -> ATNStateRef {
-        self.decision_to_state[decision]
+    pub fn get_decision_state(&self, decision: i32) -> ATNStateRef {
+        self.decision_to_state[decision as usize]
     }
 
     /// Computes the set of input symbols which could follow ATN state number
@@ -140,8 +140,8 @@ impl ATN {
     /// number {@code stateNumber}
     pub fn get_expected_tokens(
         &self,
-        state_number: isize,
-        states_stack: impl Iterator<Item = isize>, // _ctx: &Rc<Ctx::Type>,
+        state_number: i32,
+        states_stack: impl Iterator<Item = i32>, // _ctx: &Rc<Ctx::Type>,
     ) -> IntervalSet {
         let s = self.states[state_number as usize].as_ref();
         let mut following = self.next_tokens(s);
@@ -161,7 +161,7 @@ impl ATN {
             let invoking_state = self.states[state as usize].as_ref();
             let tr = invoking_state.get_transitions().first().unwrap().as_ref();
             let tr = tr.cast::<RuleTransition>();
-            following = self.next_tokens(self.states[tr.follow_state].as_ref());
+            following = self.next_tokens(self.states[tr.follow_state as usize].as_ref());
             expected.add_set(following);
             expected.remove_one(TOKEN_EPSILON);
             // ctx = c.get_parent_ctx();
